add coredata to mvvm
