+++ Alfagift-MovieApp/Movies.xcdatamodeld/Movies.xcdatamodel/contents
@@ -3,7 +3,7 @@
     <entity name="Movie" representedClassName="Movie" syncable="YES" codeGenerationType="category">
         <attribute name="adult" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="backdropPath" optional="YES" attributeType="String"/>
-        <attribute name="genreIDS" optional="YES" attributeType="Transformable" valueTransformerName="[Int]"/>
+        <attribute name="genreIDS" optional="YES" attributeType="Transformable" valueTransformerName="NSSecureUnarchiveFromData" customClassName="[Int]"/>
         <attribute name="id" optional="YES" attributeType="Integer 32" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="originalLanguage" optional="YES" attributeType="String"/>
         <attribute name="originalTitle" optional="YES" attributeType="String"/>
